+++ routes/post.js
@@ -13,15 +13,17 @@ exports.get = function(req, res){
             res.status(500);
             res.render('500');
         }else if(post){
-      	  //TODO: Get the comments associated with the post
-//          CommentModel.find({'post_id': identifier}, function(err, comment){
-//      		if(err){
-//                console.log("An error occurred when trying to retrieve comments [post-id=%s]", req.params.id);
-//            } else{
+          CommentModel.find({'post_id': identifier}).sort({date: "desc"}).exec(function(err, comments){
+      		if(err){
+                console.log("An error occurred when trying to retrieve comments [post-id=%s, error=%s]", req.params.id, err);
+                res.status(500);
+                res.render("500");
 
-      			res.render("post", {title :post.title, post: post, comments:[], admin:req.session.admin});
-//      		}
-//      	});
+            } else{
+      			comments = comments ? comments : [];
+                res.render("post", {title :post.title, post: post, comments: comments, admin:req.session.admin});
+      		}
+      	});
         
       }else{
             //Post does not exist - return 404
@@ -35,36 +37,60 @@ exports.get = function(req, res){
 
 
 exports.saveComment = function(req, res){
-    if(!req.body.comment){
-        //TODO: We should handle this error case properly when no comment is sent
-        console.log("No comment has been sent!");
-        res.redirect("/");
-    }else{
-        var commentWritten = req.body.comment;
-        var postId = mongoose.Types.ObjectId(req.params.id);
-        var commenterName = req.body.name || 'anon';
-        var title = req.params.title;
+    var postId = req.params.id;
 
-        console.log("New comment [post-id=%s, commenter-name=%s, comment=%s]", req.params.id,
-            commenterName, commentWritten);
+    if(!postId){
+        console.log("No post id provided to save commentWritten");
+        res.status(500);
+        res.render("500");
 
-        //Submitting to database
-        var newComment = CommentModel({
-            post_id: postId,
-            title: title,
-            name: commenterName,
-            comment: commentWritten
-        });
-        newComment.save(function(err){
-            if(err){
-                //TODO: Gracefully handle this error in the future
-                console.log("Unable to save new comment [post-id=%s, error=%s]", req.params.id, err)
-            }
-
-            //redirecting to homepage
-            res.redirect('/post/' + req.params.id + '/' + req.params.title);
-        });
-
-
+        return;
     }
+
+    var commentWritten = req.body.comment;
+    if(!commentWritten){
+        console.log("No comment has been sent [post-id=%s]", postId);
+        res.status(500);
+        res.render("500");
+
+        return;
+    }
+
+    var commenterName = req.body.name;
+    if(!commenterName){
+        console.log("No name was sent [post-id=%s]", postId);
+        res.status(500);
+        res.render("500");
+
+        return;
+    }
+
+
+    postId = mongoose.Types.ObjectId(postId);
+
+    console.log("New commentWritten [post-id=%s, commenter-name=%s, commentWritten=%s]", postId,
+        commenterName, commentWritten);
+
+    //Submitting to database
+    var newComment = CommentModel({
+        post_id: postId,
+        name: commenterName,
+        comment: commentWritten
+    });
+    newComment.save(function(err){
+        if(err){
+            console.log("Unable to save new commentWritten [post-id=%s, error=%s]", req.params.id, err)
+            res.status(500);
+            res.render("500");
+
+        }else{
+            //redirecting to post where the comment was just made
+            console.log("Successfully saved a comment for post [post-id=%s, commenter-name=%s]", req.params.id, commenterName)
+            res.redirect('/post/' + req.params.id + '/' + req.params.friendlyLink);
+        }
+
+    });
+
+
+
 }