+++ History.md
@@ -1,9 +1,11 @@
-1.0.0 / 2013-07-31
+1.0.0
 ====================
 * MAJOR refactoring of API (in progress).
 * Revamping of layout as well (in progress).
 * Work on Admin, new post view
 * Improved look and feel, error handling (404 + 500) and code quality
+* Readers can now submit comments to a given blog post
+* Using the configured port in blogConfig.json to bind the server to the specified port
 
 0.2.2 / 2013-02-23
 ====================